refactor(test): consolidate repo-path lookup helpers into SolutionRoot
Replaces duplicated LocateRepoFolder / LocateRepoRoot / LocateSourceRoot
helpers across architecture tests with SolutionRoot.Src / .Tests / .Path,
extracted into its own file.
